Do not delete snapshot when vm is not down

+++ src/main/java/org/ovirtChina/enginePlugin/vmBackupScheduler/bll/DeleteSnapshot.java
@@ -3,6 +3,7 @@ package org.ovirtChina.enginePlugin.vmBackupScheduler.bll;
 import java.io.IOException;
 
 import org.apache.http.client.ClientProtocolException;
+import org.ovirt.engine.sdk.decorators.VM;
 import org.ovirt.engine.sdk.exceptions.ServerException;
 import org.ovirtChina.enginePlugin.vmBackupScheduler.common.Task;
 import org.ovirtChina.enginePlugin.vmBackupScheduler.common.TaskType;
@@ -16,9 +17,14 @@ public class DeleteSnapshot extends DeleteSDKTask {
 
     @Override
     protected void deleteTask(Task task) throws ClientProtocolException, ServerException, IOException {
-        api.getVMs().get(task.getVmID()).getSnapshots().getById(task.getBackupName()).delete();
-        DbFacade.getInstance().getTaskDAO().delete(task.getVmID(), task.getBackupName());
-        log.info("delete snapshot: " + task.getBackupName() + " for vm: " + task.getVmID() + " has initiated.");
+        VM vm = api.getVMs().get(task.getVmID());
+        if (vm.getStatus().getState().equals("down")) {
+            vm.getSnapshots().getById(task.getBackupName()).delete();
+            DbFacade.getInstance().getTaskDAO().delete(task.getVmID(), task.getBackupName());
+            log.info("delete snapshot: " + task.getBackupName() + " for vm: " + task.getVmID() + " has initiated.");
+        } else {
+            log.debug("cancel deletion of snapshot: " + task.getBackupName() + " for vm: " + task.getVmID() + ", because vm is not down.");
+        }
     }
 
 }
